ref(seer): Move feature-run dispatch into SeerAgentClient
Consolidate the feature-run API onto SeerAgentClient.start_feature_run
(companion to start_run) instead of a standalone feature_run module, so there's
a single, consistent entry point. Night shift constructs the client like any
other caller, which means its manual triggers now run through the standard Seer
access gate (gen-ai-features / hide_ai_features) and skip dispatch for orgs
without access.

Also drops some over-verbose comments.
